+++ lib/src/saslprep.dart
@@ -1,4 +1,6 @@
-import "package:unorm_dart/unorm_dart.dart" as unorm;
+import 'dart:typed_data';
+
+import 'package:unorm_dart/unorm_dart.dart' as unorm;
 
 import 'code_points/bidirectional_l.dart';
 import 'code_points/bidirectional_r_al.dart';
@@ -12,38 +14,38 @@ class Saslprep {
   /// Convert provided string into an array of Unicode Code Points.
   /// Based on https://stackoverflow.com/a/21409165/1556249
   /// and https://www.npmjs.com/package/code-point-at.
-  static List<int> toCodePoints(String input) {
-    List<int> codePoints = [];
-    int size = input.length;
-
-    for (int i = 0; i < size; i += 1) {
+  static Uint32List toCodePoints(String input) {
+    var codePoints = Uint32List(input.length);
+    var sizeOffset = 0;
+    var codePointIndex = 0;
+    for (var i = 0; i < input.length; i++) {
       var before = input.codeUnitAt(i);
 
-      if (before >= 0xd800 && before <= 0xdbff && size > i + 1) {
+      if (before >= 0xd800 && before <= 0xdbff && input.length > i + 1) {
         var next = input.codeUnitAt(i + 1);
 
         if (next >= 0xdc00 && next <= 0xdfff) {
-          codePoints.add((before - 0xd800) * 0x400 + next - 0xdc00 + 0x10000);
-          i += 1;
+          codePoints[codePointIndex++] = ((before - 0xd800) * 0x400 + next - 0xdc00 + 0x10000);
+          i++;
+          sizeOffset++;
           continue;
         }
       }
-
-      codePoints.add(before);
+      codePoints[codePointIndex++] = (before);
     }
-
+    codePoints = codePoints.sublist(0, codePoints.length - sizeOffset);
     return codePoints;
   }
 
   /// This computes the saslprep algorithm. to allow allow unassigned use the
   /// [options] and set [options.allowUnassigned] to true
-  static saslprep(String input, {SaslprepOptions options}) {
+  static String saslprep(String input, {SaslprepOptions options}) {
     if (input.isEmpty) {
       return '';
     }
 
     // 1. Map
-    Iterable<int> mapped_input = toCodePoints(input)
+    var mapped_input = toCodePoints(input)
         // 1.1 mapping to space
         .map((character) => (non_ASCII_space_characters.contains(character)
             ? 0x20
@@ -52,12 +54,12 @@ class Saslprep {
         .where((character) => !commonly_mapped_to_nothing.contains(character));
 
     // 2. Normalize
-    String normalized_input = unorm.nfkc(String.fromCharCodes(mapped_input));
+    var normalized_input = unorm.nfkc(String.fromCharCodes(mapped_input));
 
-    List<int> normalized_map = toCodePoints(normalized_input);
+    var normalized_map = toCodePoints(normalized_input);
 
     // 3. Prohibit
-    bool hasProhibited = normalized_map
+    var hasProhibited = normalized_map
         .any((character) => prohibited_characters.contains(character));
 
     if (hasProhibited) {
@@ -67,7 +69,7 @@ class Saslprep {
 
     // Unassigned Code Points
     if (options == null || options.allowUnassigned != true) {
-      bool hasUnassigned = normalized_map
+      var hasUnassigned = normalized_map
           .any((character) => unassigned_code_points.contains(character));
       if (hasUnassigned) {
         throw Exception(
@@ -76,9 +78,9 @@ class Saslprep {
     }
 
     // 4. check bidi
-    bool hasBidiRAL = normalized_map
+    var hasBidiRAL = normalized_map
         .any((character) => bidirectional_r_al.contains(character));
-    bool hasBidiL = normalized_map
+    var hasBidiL = normalized_map
         .any((character) => bidirectional_l.contains(character));
 
     // 4.1 If a string contains any RandALCat character, the string MUST NOT
@@ -91,9 +93,9 @@ class Saslprep {
     //4.2 If a string contains any RandALCat character, a RandALCat
     //character MUST be the first character of the string, and a
     //RandALCat character MUST be the last character of the string.
-    bool isFirstBidiRAL =
+    var isFirstBidiRAL =
         bidirectional_r_al.contains(normalized_input.codeUnitAt(0));
-    bool isLastBidiRAL = bidirectional_r_al.contains(
+    var isLastBidiRAL = bidirectional_r_al.contains(
             normalized_input.codeUnitAt(normalized_input.length - 1));
 
     if (hasBidiRAL && !(isFirstBidiRAL && isLastBidiRAL)) {