+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 1.0.1
+
+- update dependencies
+- add pedantic package
+- replaced int list by Uint32List for better performance
+
 ### 1.0.0
 
 - initial deployment to https://pub.dev